Moves ThisInvokeContext to the program-runtime crate.
