+++ cli/Cargo.toml
@@ -35,9 +35,9 @@ solana-client = { path = "../client", version = "=1.9.0" }
 solana-config-program = { path = "../programs/config", version = "=1.9.0" }
 solana-faucet = { path = "../faucet", version = "=1.9.0" }
 solana-logger = { path = "../logger", version = "=1.9.0" }
+solana-program-runtime = { path = "../program-runtime", version = "=1.9.0" }
 solana_rbpf = "=0.2.14"
 solana-remote-wallet = { path = "../remote-wallet", version = "=1.9.0" }
-solana-runtime = { path = "../runtime", version = "=1.9.0" }
 solana-sdk = { path = "../sdk", version = "=1.9.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.9.0" }
 solana-version = { path = "../version", version = "=1.9.0" }
