+++ program-runtime/src/log_collector.rs
