+++ programs/config/Cargo.toml
@@ -18,7 +18,7 @@ solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 
 [dev-dependencies]
 solana-logger = { path = "../../logger", version = "=1.9.0" }
-solana-runtime = { path = "../../runtime", version = "=1.9.0" }
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 
 [lib]
 crate-type = ["lib"]
