+++ programs/stake/Cargo.toml
@@ -20,6 +20,7 @@ solana-frozen-abi = { path = "../../frozen-abi", version = "=1.9.0" }
 solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.9.0" }
 solana-metrics = { path = "../../metrics", version = "=1.9.0" }
 solana-sdk = { path = "../../sdk", version = "=1.9.0" }
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 solana-vote-program = { path = "../vote", version = "=1.9.0" }
 solana-config-program = { path = "../config", version = "=1.9.0" }
 thiserror = "1.0"
@@ -27,7 +28,6 @@ thiserror = "1.0"
 [dev-dependencies]
 proptest = "1.0"
 solana-logger = { path = "../../logger", version = "=1.9.0" }
-solana-runtime = { path = "../../runtime", version = "=1.9.0" }
 
 [build-dependencies]
 rustc_version = "0.4"
