+++ programs/vote/Cargo.toml
@@ -24,7 +24,7 @@ solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 thiserror = "1.0"
 
 [dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "=1.9.0" }
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 
 [build-dependencies]
 rustc_version = "0.4"
