+++ rbpf-cli/Cargo.toml
@@ -14,7 +14,7 @@ serde = "1.0.130"
 serde_json = "1.0.68"
 solana-bpf-loader-program = { path = "../programs/bpf_loader", version = "=1.9.0" }
 solana-logger = { path = "../logger", version = "=1.9.0" }
+solana-program-runtime = { path = "../program-runtime", version = "=1.9.0" }
 solana-sdk = { path = "../sdk", version = "=1.9.0" }
 solana_rbpf = "=0.2.14"
-solana-runtime = { path = "../runtime", version = "=1.9.0" }
 time = "0.3.4"
